refactor: 테이블 관련 코드를 connection에서 table로 분리
